requirements-dev.txt
-r requirements.txt
black
codespell
juju
ops-scenario
pyright
pytest
pytest-operator
ruff
